🔨 .env 환경 변수 설정 및 docker-compose 작성

+++ src/main/java/Backend/HIFI/user/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 public interface UserRepository extends JpaRepository<User, Integer> {
 
     User findUserByEmail(String email);
-    User findUserById(int id);
+    User findUserById(Long id);
 
 
 //    @Modifying
